Reorganize file structure and refactor code
In addition to reorganizing the app's file structure, this commit also refactors code by using namespace imports and reducing dependency through functions and modules. It also implements an empty UI for the app's default setup as opposed to containing pre-populated projects and tasks.

This commit also validates and adds projects via the 'Add project' button as well as implement a live date feature for both the home page and future implementations.