Synchronize status of same task across sections
This commit updates the status of a task across all sections listed in the sidebar. If a task was checked or unchecked prior to this commit, only the displayed HTML elements of the task would update. If the task belonged to one of the non-displayed 'Upcoming' sections, it would not update. This commit fixes the issue by targeting hidden elements and updating them accrodingly.