Fix 'sort by' button bug
This commit fixes the bug in which the background color of the first 'sort by' button's selected filter is removed and multiple filters are highlighted of the 'sort by' button clicked on. This was because a query selector was targeting the selected filter of the first 'sort by' button present on the DOM as opposed to the 'sort by' button that was clicked on.